+++ packages/utils/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 22.2.0 (2026-08-30)
+
+### 🚀 Features
+
+- **eslint-plugin:** add allowExplicitOnPush option, auto-remove explicit OnPush when false ([#3153](https://github.com/angular-eslint/angular-eslint/pull/3153))
+
+### ❤️ Thank You
+
+- Evgeny Stepanovych @undsoft
+
 ## 22.1.0 (2026-07-12)
 
 ### 🚀 Features